fix rscolor bug when alpha is negative number

+++ rosen/modules/render_service_base/src/common/rs_color.cpp
@@ -78,8 +78,10 @@ RSColor RSColor::operator/(float scale) const
 
 uint32_t RSColor::AsRgbaInt() const
 {
-    return (std::clamp<uint32_t>(alpha_, 0, UINT8_MAX)) | (std::clamp<uint32_t>(red_, 0, UINT8_MAX) << 24) |
-           (std::clamp<uint32_t>(green_, 0, UINT8_MAX) << 16) | (std::clamp<uint32_t>(blue_, 0, UINT8_MAX) << 8);
+    return (static_cast<uint32_t>(std::clamp<int16_t>(alpha_, 0, UINT8_MAX))) |
+           (static_cast<uint32_t>(std::clamp<int16_t>(red_, 0, UINT8_MAX) << 24)) |
+           (static_cast<uint32_t>(std::clamp<int16_t>(green_, 0, UINT8_MAX) << 16)) |
+           (static_cast<uint32_t>(std::clamp<int16_t>(blue_, 0, UINT8_MAX) << 8));
 }
 
 RSColor RSColor::FromRgbaInt(uint32_t rgba)
@@ -89,8 +91,10 @@ RSColor RSColor::FromRgbaInt(uint32_t rgba)
 
 uint32_t RSColor::AsArgbInt() const
 {
-    return ((std::clamp<uint32_t>(alpha_, 0, UINT8_MAX)) << 24) | ((std::clamp<uint32_t>(red_, 0, UINT8_MAX)) << 16) |
-           ((std::clamp<uint32_t>(green_, 0, UINT8_MAX)) << 8) | (std::clamp<uint32_t>(blue_, 0, UINT8_MAX));
+    return ((static_cast<uint32_t>(std::clamp<int16_t>(alpha_, 0, UINT8_MAX))) << 24) |
+           ((static_cast<uint32_t>(std::clamp<int16_t>(red_, 0, UINT8_MAX))) << 16) |
+           ((static_cast<uint32_t>(std::clamp<int16_t>(green_, 0, UINT8_MAX))) << 8) |
+           (static_cast<uint32_t>(std::clamp<int16_t>(blue_, 0, UINT8_MAX)));
 }
 
 RSColor RSColor::FromArgbInt(uint32_t argb)
@@ -101,8 +105,10 @@ RSColor RSColor::FromArgbInt(uint32_t argb)
 
 uint32_t RSColor::AsBgraInt() const
 {
-    return (std::clamp<uint32_t>(alpha_, 0, UINT8_MAX)) | ((std::clamp<uint32_t>(red_, 0, UINT8_MAX)) << 8) |
-           ((std::clamp<uint32_t>(green_, 0, UINT8_MAX)) << 16) | ((std::clamp<uint32_t>(blue_, 0, UINT8_MAX)) << 24);
+    return (static_cast<uint32_t>(std::clamp<int16_t>(alpha_, 0, UINT8_MAX))) |
+           ((static_cast<uint32_t>(std::clamp<int16_t>(red_, 0, UINT8_MAX))) << 8) |
+           ((static_cast<uint32_t>(std::clamp<int16_t>(green_, 0, UINT8_MAX))) << 16) |
+           ((static_cast<uint32_t>(std::clamp<int16_t>(blue_, 0, UINT8_MAX))) << 24);
 }
 
 RSColor RSColor::FromBgraInt(uint32_t bgra)
